Festitivities test
This is the final version for this tech test. It has  the .git folder.
However, in an alternative branch, it's going to be push some classes developed for ETL, but for time, I couldn't fix the error between hibernate and joda.time with the deserealization in the deadline. I hope fix the problems and deliver a new version for my own purpose.